Refactor normalize_density function to use np.float32 instead of np.float16 for improved precision

+++ hotspot2/utils.py
@@ -86,7 +86,7 @@ def ensure_contig_exists(func):
     return wrapper
 
 def normalize_density(density, total_cutcounts):
-    return (density / total_cutcounts * 1_000_000).astype(np.float16)
+    return (density / total_cutcounts * 1_000_000).astype(np.float32)
 
 def run_bam2_bed(bam_path, tabix_bed_path):
     script_dir = os.path.dirname(os.path.abspath(__file__))
